Adds an interface for generating the upm package
